Nearly Done
I just have to finish the level order traversal and I will be ready for
testing.

+++ Makefile
@@ -4,17 +4,17 @@
 #
 #   This program is entirely my own work
 CC=gcc
-CFLAGS=-Wall -g -std=c99 -c
-OBJECTS=scanner.o node.o Fatal.o vertex.o cdll.o binheap.o graph.o
+CFLAGS=-Wall -Wextra -g -std=c99 -c
+OBJECTS=scanner.o node.o Fatal.o vertex.o cdll.o binheap.o graph.o queue.o
 TESTOBJECTS=scanner.o node.o Fatal.o vertex.o cdll.o binheap.o graph.o
 
 all: prim
 
 prim: prim.o ${OBJECTS}
-	${CC} -Wall -std=c99 -g -o prim prim.o ${OBJECTS} -lm
+	${CC} -Wall -Wextra -std=c99 -g -o prim prim.o ${OBJECTS} -lm
 
 test: prim.o testing.o ${TESTOBJECTS}
-	${CC} -Wall -std=c99 -g -o testing testing.o ${TESTOBJECTS} -lm
+	${CC} -Wall -Wextra -std=c99 -g -o testing testing.o ${TESTOBJECTS} -lm
 	./testing data1
 
 testing.o: testing.c cdll.h vertex.h node.h binheap.h scanner.h graph.h Fatal.h
@@ -38,6 +38,9 @@ binheap.o: binheap.c binheap.h node.h Fatal.h cdll.h
 graph.o: graph.c graph.h vertex.h Fatal.h
 	${CC} ${CFLAGS} graph.c
 
+queue.o: queue.c queue.h node.h Fatal.h
+	${CC} ${CFLAGS} queue.c
+
 Fatal.o: Fatal.c Fatal.h
 	${CC} ${CFLAGS} Fatal.c
 
@@ -45,7 +48,6 @@ scanner.o: scanner.c scanner.h
 	${CC} ${CFLAGS} scanner.c
 
 clean:
-	rm ${OBJECTS} prim
-	if [ -f "testing" ]; then
-   		rm testing
-	fi
+	rm ${OBJECTS}
+	if [ -f "testing" ]; then rm testing; fi
+	if [ -f "prim" ]; then rm prim; fi
